requirements.txt
Django>=5.1.6
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
gunicorn>=21.2.0
djangorestframework>=3.15.2 